% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ggamma_model}
\alias{ggamma_model}
\title{Generalised gamma impedance distribution model}
\format{
A named list defining the model form with components:
\describe{
  \item{name}{Character string identifying the model.}
  \item{parameters}{Character vector of parameter names.}
  \item{log_density}{Function computing the log-density.}
  \item{gradient}{Named list of score functions.}
  \item{hessian}{Named list of second derivatives.}
}
}
\usage{
ggamma_model
}
\description{
A flexible impedance model based on the generalised gamma distribution.
This model is used to represent marginal impedance distributions of travel
times and decomposes naturally into:
}
\details{
* an impedance function: \eqn{f(x) = \exp\{-(\lambda x)^r\}}
* a spatial weight function: \eqn{\g(x) = x^{\alpha - 1}}

As such it assumes that the impedance function is in power exponential form,
and that the spatial weight grows as a power function of impedance.

The full density is:
\deqn{
\p(x) =
  r \frac{\lambda^\alpha}{\Gamma(\alpha / r)}
  x^{\alpha - 1} \exp\{-(\lambda x)^r\}
}


Parameter interpretation:
\itemize{
  \item \code{r}: functional form of impedance (shape).
  \item \code{alpha}: effective spatial dimension
  \item \code{lambda}: rate parameter controlling travel-time decay.
}
}
\keyword{internal}
