% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmdc.R
\name{wmdc}
\alias{wmdc}
\title{Calibrate an impedance function using a marginal impedance distribution}
\usage{
wmdc(
  data,
  shape = 1,
  dimension = 2,
  free_shape = FALSE,
  free_dimension = FALSE,
  maxLik_output = FALSE
)
}
\arguments{
\item{data}{An observed marginal impedance distribution in one of the 
following two forms:
\itemize{
   \item A data frame containing a frequency table for the impedance
variable. The data frame should have two columns: the first column contains
the impedance values (as individual numeric values, not as ranges), and the
second column contains the corresponding frequencies.
\item A vector containing numeric impedance values
}}

\item{shape}{A numeric value representing the fixed or starting value for the
shape parameter r in the power exponential form. A shape parameter of 1
results in an exponential form, while one of 2 results in a Gaussian form
The default value is 1, i.e. exponential.}

\item{dimension}{A numeric value representing the fixed effective spatial
dimension of the study environment, or the starting value for the dimension
parameter when `free_dimension = TRUE`. Expected range is [1,2], including
non-integer values, as networks have a fractal structure spanning a subset
the 2-D plane. Default is 2, reflecting free movement across 2-D space.}

\item{free_shape}{A Boolean value indicating whether to allow the shape to be
a free parameter in the calibration process. If TRUE, the shape parameter
will be estimated from the data rather than being fixed at the value
indicated by `shape`. Default is FALSE.}

\item{free_dimension}{A Boolean value indicating whether to allow the spatial
dimension to be a free parameter in the calibration process. If TRUE, the
spatial dimension will be estimated from the data rather than being fixed at
the value indicated by `dimension`. Default is FALSE.}

\item{maxLik_output}{A Boolean value indicating whether the full output from
the maxLik operation should be returned in the output list}
}
\value{
A list containing the following elements:
\itemize{
   \item \code{impedance_function}: A function that computes the calibrated
   impedance decay for a given impedance value, based on the specified form
   and calibrated parameters.
   \item \code{weight_function}: A weight function capturing the 
   distortion induced by systematic variation in the spatial structure of 
   origins and destinations across impedance.
   \item \code{density_function}: The probability density function for the
   marginal distribution of impedance, based on the calibrated parameters and
   specified form.
   \item \code{survival_function}: The survival function (P(X>=x)) for the
   marginal distribution of impedance, based on the calibrated parameters and
   specified form.
   \item \code{parameters}: A list of the calibrated parameters of the
   impedance function.
   \item \code{parameters_se}: A list of standard errors of the calibrated
   parameters of the impedance function.
   \item \code{dimension}: The effective spatial dimension used in the
   calibration, either fixed or estimated.
   \item \code{dimension_se}: The standard error of the estimated spatial
   dimension, if it was estimated; otherwise NA.
   \item \code{log_likelihood}: The log-likelihood value of the calibrated 
   model, indicating the goodness of fit of the model to the data.
   \item \code{AIC}: The Akaike Information Criterion for the fitted model
   indicating the relative quality of the model fit, with lower values
   indicating a better fit relative to other models.
   \item \code{survival_R_sq}: The R-squared value for the fit of the
   estimated survival function to the observed survival function, indicating
   how well the calibrated model captures the distribution of impedance in
   the data. This is not recommended for model selection but can be used as a
   descriptive measure of fit, and aids comparison with the commonly used
   method of fitting impedance functions to the observed survival function of
   impedance values.
   \item \code{maxLik_output}: (optional) The full output from the maxLik
   operation, including parameter estimates, standard errors, convergence
   information, and more. This is included in the output list if
   `maxLik_output`is set to TRUE, and is useful for users who want to inspect
   the details of the optimization process or perform additional analyses
   based on the maxLik results.
}
}
\description{
This function calibrates an impedance function using the marginal
distribution of an impedance variable, expressed as a frequency table or
vector of observations.
}
